Rename ports packages
