+++ domain/src/main/kotlin/fr/sacane/jmanager/domain/port/api/UserRegister.kt
@@ -1,11 +1,11 @@
-package fr.sacane.jmanager.domain.port.apiside
+package fr.sacane.jmanager.domain.port.api
 
 import fr.sacane.jmanager.domain.models.Response
 import fr.sacane.jmanager.domain.models.User
-import fr.sacane.jmanager.domain.hexadoc.PortToLeft
-import fr.sacane.jmanager.domain.port.serverside.UserTransaction
+import fr.sacane.jmanager.domain.hexadoc.LeftPort
+import fr.sacane.jmanager.domain.port.spi.UserTransaction
 
-@PortToLeft
+@LeftPort
 class UserRegister(private val userPort: UserTransaction) {
     fun signIn(user: User): Response<User> {
         val userResponse = userPort.save(user) ?: return Response.invalid()
