+++ domain/src/main/kotlin/fr/sacane/jmanager/domain/port/spi/LoginInventory.kt
@@ -1,12 +1,12 @@
-package fr.sacane.jmanager.domain.port.serverside
+package fr.sacane.jmanager.domain.port.spi
 
-import fr.sacane.jmanager.domain.hexadoc.PortToRight
+import fr.sacane.jmanager.domain.hexadoc.RightPort
 import fr.sacane.jmanager.domain.models.Password
 import fr.sacane.jmanager.domain.models.Ticket
 import fr.sacane.jmanager.domain.models.Token
 import fr.sacane.jmanager.domain.models.UserId
 
-@PortToRight
+@RightPort
 interface LoginInventory {
     fun login(userPseudonym: String, password: Password): Ticket?
     fun logout(userId: UserId, token: Token): Ticket?
