+++ infra/src/main/kotlin/fr/sacane/jmanager/JmanagerConfiguration.kt
@@ -1,11 +1,11 @@
 package fr.sacane.jmanager
 
-import fr.sacane.jmanager.domain.port.apiside.LoginManager
-import fr.sacane.jmanager.domain.port.apiside.TransactionReaderAdapter
-import fr.sacane.jmanager.domain.port.apiside.UserRegister
-import fr.sacane.jmanager.domain.port.serverside.LoginInventory
-import fr.sacane.jmanager.domain.port.serverside.TransactionRegister
-import fr.sacane.jmanager.domain.port.serverside.UserTransaction
+import fr.sacane.jmanager.domain.port.api.LoginManager
+import fr.sacane.jmanager.domain.port.api.TransactionReaderAdapter
+import fr.sacane.jmanager.domain.port.api.UserRegister
+import fr.sacane.jmanager.domain.port.spi.LoginInventory
+import fr.sacane.jmanager.domain.port.spi.TransactionRegister
+import fr.sacane.jmanager.domain.port.spi.UserTransaction
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
